requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pillow>=9.0.0
beautifulsoup4>=4.12.0